Clean up RMQ initialization
Added several options to the RabbitMQ section of the config file and
updated the rmq_init() function to use all of them.  Also removed a
bunch of the RMQ parameters from the global class because they only
needed to be local to rmq_init().

ToDo: oddsub.py is probably going to need exactly the same rmq_init()
function.  Probably ought to move it to its own python module.