Changing project layout
Changing the layout so that most files are in an oddmon package.  (This
keeps us from polluting the site-packages directory with individual .py
files.)  Updated the setup.py and monctl.py files to reflect the change.

ToDo: The upstart .conf files probably need to be updated.  Also,
there's a bit of a conumdrum with lfs_utils.  It's imported by the
plugins, but because of how the plugins themselves are imported, it has
to be in the oddmon directory, not oddmon/metric_plugins...