Change packaging of config file
It's best if the RPM package doesn't overwrite the local config file on
an update.  Change setup.py to include oddmon.cfg.sample instead of
oddmon.cfg and include a post_install script that checks to see if
oddmon.cfg already exists.  (This is a work-around for the fact that
python distutils doesn't support RPM's %config(noreplace) option.)