Replace 'host' and 'sourcetype' fields
The Splunk server treats 'host' and 'sourcetype' fields specially.  More
to the point, it is automatically adding them to the events when it
reads in the brw_stats and job_stats log files.  The fact that we had
already specified them in the text seems to be ignored...  So, replace
'host' with 'OST' and sourcetype with 'datatype' in the text the metrics
write.