Stagger connect attempts so we don't flood the RMQ server
Add a small ( < 1 second) random delay before attempting to connect to the RMQ server.  (This is mostly for use in cases where someone uses PDSH (or simliar) to start up hundreds of these clients simultaneously.)  Also add some code to deal with a timeout exception when connecting: it waits a random amount of time (between 1 and 5 seconds) and tries again.