Add a "drain mode" to the aggregator
Add a --drain command line option.  When specified, the aggregator will connect to the RMQ server and retrieve messages, but then throw them out.  They will not be passed to the handler functions of the plugins, and thus not sent to the Splunk server.  Useful for occasions when we don't want to push data into Splunk, but also don't want the messages filling up the queue on the RMQ servers.