Overhaul the job_stats handling code
The job stats code was completely mis-handling the timestamps and thus
not outputting proper data.  It took a fairly extensive overhaul to
fix this.  Timestamp comparisons are now handled solely by the
collector.  This saves resources on the RabbitMQ servers by not sending
messages that the aggregator will then just discard.  Also added a
bunch of comments since the code was nearly devoid of them.