Clean up dictionary handling
If we want to iterate through a dictionary, it's more efficient to use
"for k in d:"  than "for k in d.keys():".  There's place where we do
actually modify a dictionary as we iterate through it, and for that
we use "for k in list(d):".  (In Python2, this is effectively the same as
"for k in d.keys():", but in Python3, d.keys() doesn't actually return a
separate list object.)