Change file handlers for Splunk log files
Switch from RotatingFileHandler to TimedRotatingFileHandler for the
logger objects that output the Splunk data.  (Switching because the
timed version appends a time and data to the filename when it rotates
and that makes it easier to tell them apart.  RotatingFileHandler just
appends .1, .2, .3, etc...).  Also changing the specs to keep 30 hours
of files.