Support Lustre 2.8 job_stats files
Re-factor the read_target_stats() function to handle Lustre 2.8 job_stats files.
Simplified it a little, too.  It's less picky about what it expects to find
in the job_stats files.  (This is a good new/bad news situation. The good news is
the same function can handle different versions of luster.  The bad news is that
it's less likely to flag bogus or corrupted data.)