Don't try to process an empty job_stats file
If we scan the job_stats file and it's empty (for instance, because
we're running on my test system and I haven't turned job_stats on in
Lustre), then skip the remaining processing.  This avoids having the
plugin throw an exception when its sanity checks for the required fields
('snapshot_time' and 'job_id') fail.