Subscriber can now log error messages from the publishers
If the subscriber receives some decoded JSON who's key is 'client_msg'
(rather than the normal name of the plugin), it will log all the data
associated with that key as errors.  This allows the publishers to send
error messages to a place where they might actually be seen.  (Nobody is
regularly checking the syslog for the servers running the publishers.)