Publisher sanity checks its output buffer
Publisher now runs a sanity check (basically, just a JSON decode operation)
on its output buffer before sending the data to the RMQ server.  If a problem
is detected, it sends a separate message to the publisher using the mechanism
described in the previous commit.

Rationale: Since the last change to the production servers, we've been seeing
occasional cases where the subscriber receives invalid JSON data from the
RMQ server.  Initial debugging appears to indicate the problem is with RabbitMQ
(either the servers themselves, or the Pika library I'm using to communicate
with them).  Once this change is deployed to production, I'll be able to tell
if the errors are originating in my code or not.