Don't use persistant connections for the publisher
We've noticed that the publisher seem to be crashing with 'connection
reset' errors fairly frequently.  It's not clear why, but it appears the
pika.BlockingConnection class doesn't like long-term connections.
(Interestly, the subscriber uses pika.SelectConnection and it doesn't
suffer from these problems.)

The workaround is to open a new connection for each message and then close
it once the message has been sent.  Since the production servers are
configured to only send one (fairly large) message every 5 minutes,
creating a new connection each time isn't a lot of extra overhead.