Execute all pika calls in a subprocess
We were having troubles with the collectors hanging when they tried to
close the connection to the RMQ server.  It looks like they would send a
close request, and then hang waiting for a reply.  (Why they never
received the reply is currently unknown.  Can't tell if the server actually
sent it or not...)

This change moves all the Pika calls into a sub-process.  That way, if the
close call hangs, the main process can terminate the subprocess.

(Also bumped the version to 1.6.1.)