Fix negative values in job_stats
Fix the longstanding bug that causes negative values in the job_stats
data.  See the comments at the bottom of read_target_stats() for a
complete explanation, but the short version is that if lustre removes a
job from the job_stats file, then we need to make sure we purge it from
our internal data structures, too.

Note that the fact that we weren't purging these jobs was also
effectively a memory leak (albeit a fairly slow one).